💩: 单元测试修改

+++ src/main/resources/schema.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS `customer` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE IF NOT EXISTS `tb_account`(
-    `id`        INTEGER PRIMARY KEY AUTO_INCREMENT,
+    `id`        BIGINT PRIMARY KEY AUTO_INCREMENT,
     `user_name` VARCHAR(100),
     `age`       INTEGER,
     `birthday` DATETIME,
